- Added support for invalid SSL certificates, - Added thumbnails to external links, - Added support for PDF external links, - Added main temporary method to harvest some external links, - Added search in the document title (with a 2x ranking boost).

+++ src/main/java/au/gov/aims/eatlas/searchengine/rest/Index.java
@@ -37,7 +37,7 @@ import java.util.List;
 public class Index {
     private static final Logger LOGGER = Logger.getLogger(Index.class.getName());
 
-    // TODO Require login!
+    // TODO Require login?
     @GET
     @Path("status")
     @Produces(MediaType.APPLICATION_JSON)
@@ -51,8 +51,8 @@ public class Index {
         }
 
         // TODO Implement
-        long remaining = 5;
-        long total = 1024;
+        int remaining = 5;
+        int total = 1024;
 
         JSONObject jsonStatus = new JSONObject()
             .put("remaining", remaining)
